SignupLinkPresenter now handles signup links
Moved methods from FinderPresenter and ResultSetPresenter
that relate to signup links to its own model
